Implemented the first version of both html and css
Incredibly large commit. I kind of forgot about git completely, also
because I got frustrated with this exercise. I've mostly implemented
everything that's required for this assignment. I just want to personalize
it a little more and also work out a couple of kinks.

I tried to make a lot of styles as modular as possible, but in some cases
that's not really possible. That's fine, but in the first version (which
isn't committed to git at all) had a lot of hard-coded styles per section.
Right now there's some hard-coded stuff, but I tried to make the amount
at a minimum. For the most part I tried to make a lot of re-used styles into
their own little classes.